refactor: auth handlers
