+++ src/shared/api/hooks/index.ts
@@ -2,4 +2,3 @@ export * from './useAuth';
 export * from './useDeleteReview';
 export * from './useToggleCharacteristic';
 export * from './useToggleFavorite';
-export * from './useAuthHandlers';
